add comments in refvar transformer Refactor change
