+++ pkg/transformers/refvars.go
@@ -13,6 +13,7 @@ type refvarTransformer struct {
 	vars        map[string]string
 }
 
+// NewRefVarTransformer returns a Trasformer that replaces $(VAR) style variables with values.
 func NewRefVarTransformer(vars map[string]string) (Transformer, error) {
 	return &refvarTransformer{
 		vars: vars,
@@ -33,17 +34,16 @@ func NewRefVarTransformer(vars map[string]string) (Transformer, error) {
 	}, nil
 }
 
+// Transform determines the final values of variables:
+//
+// 1.  Determine the final value of each variable:
+//     a.  If the variable's Value is set, expand the `$(var)` references to other
+//         variables in the .Value field; the sources of variables are the declared
+//         variables of the container and the service environment variables
+//     b.  If a source is defined for an environment variable, resolve the source
+// 2.  Create the container's environment in the order variables are declared
+// 3.  Add remaining service environment vars
 func (rv *refvarTransformer) Transform(resources resmap.ResMap) error {
-	// Determine the final values of variables:
-	//
-	// 1.  Determine the final value of each variable:
-	//     a.  If the variable's Value is set, expand the `$(var)` references to other
-	//         variables in the .Value field; the sources of variables are the declared
-	//         variables of the container and the service environment variables
-	//     b.  If a source is defined for an environment variable, resolve the source
-	// 2.  Create the container's environment in the order variables are declared
-	// 3.  Add remaining service environment vars
-
 	for GVKn := range resources {
 		obj := resources[GVKn].Unstruct()
 		objMap := obj.UnstructuredContent()
